Restructure e2e test suite for better run times
