+++ e2e/fixtures/status.go
@@ -34,7 +34,6 @@ import (
 
 	fdbv1beta2 "github.com/FoundationDB/fdb-kubernetes-operator/v2/api/v1beta2"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/wait"
 )
 
 // getStatusFromOperatorPod returns fdb status queried through the operator Pod.
@@ -222,16 +221,10 @@ func (fdbCluster *FdbCluster) IsAvailable() bool {
 }
 
 // WaitUntilAvailable waits until the cluster is available.
-func (fdbCluster *FdbCluster) WaitUntilAvailable() error {
-	return wait.PollUntilContextTimeout(
-		context.Background(),
-		100*time.Millisecond,
-		60*time.Second,
-		true,
-		func(_ context.Context) (bool, error) {
-			return fdbCluster.IsAvailable(), nil
-		},
-	)
+func (fdbCluster *FdbCluster) WaitUntilAvailable() {
+	gomega.Eventually(func() bool {
+		return fdbCluster.IsAvailable()
+	}).To(gomega.BeTrue())
 }
 
 // StatusInvariantChecker provides a way to check an invariant for the cluster status.
